cortex-m4: move hardfault handler to arch
So far it has just been copied a couple times.
